feature(UI): restructure project packages and implement ClientsListScreen
- Relocate `OrderListScreen`, `ClientsListScreen`, and `Components` to dedicated `ui` subpackages.
- Update `MainActivity` to display `ClientsListScreen`.
- Add `material-icons-extended` dependency and update `libs.versions.toml`.
- Initialize placeholder `OrderDatabase` in `data` package.
- Extract `BottomNavigationBar` into a reusable component.
