+++ app/src/main/java/com/example/ordermanagementcake/ui/components/Components.kt
@@ -1,4 +1,4 @@
-package com.example.ordermanagementcake.ui.theme
+package com.example.ordermanagementcake.ui.components
 
 
 import androidx.compose.material.icons.Icons
